requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
chromadb
pypdf
python-dotenv
streamlit